+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `semihosting` feature: enables clean QEMU termination via semihosting `SYS_EXIT` calls, preventing hanging processes when running `cargo test` under QEMU
+
 ## [0.1.1] - 2026-02-18
 
 ### Fixed
@@ -21,5 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UART driver
 - Optional partial `std` support
 
+[Unreleased]: https://github.com/AdaCore/rust-zynqmp/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/AdaCore/rust-zynqmp/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/AdaCore/rust-zynqmp/releases/tag/v0.1.0